Base Users completed for eventhandling
