+++ src/Phase1/Users/SwipeUser.java
@@ -1,5 +1,6 @@
 package Phase1.Users;
 
+import javax.swing.text.html.ImageView;
 import java.util.ArrayList;
 import java.util.Date;
 /**
@@ -11,6 +12,8 @@ public class SwipeUser extends User {
 
     private String gender;
     private String bio;
+    private String image;
+    private int age;
 
 
     public SwipeUser(int id) {
@@ -27,11 +30,38 @@ public class SwipeUser extends User {
 
     }
 
+    public ImageView getImage(){
+        try{
+            return new ImageView(new Image(new FileInputStream(this.image)));}
+        catch(Exception io){
+            return new ImageView();
+        }
+    }
+
+    public boolean setImage(String path){
+        try{
+            FileInputStream f = new FileInputStream(path);
+            this.image = path;
+            return true;
+        }
+        catch(Exception e){
+            return false;
+        }
+    }
+
     public String getGender() {
         return gender;
     }
 
     public void setGender(String gender) {
         this.gender = gender;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
     }
 }
