+++ src/Phase1/Users/User.java
@@ -10,8 +10,7 @@ public abstract class User {
     private int id;
     private String fName;
     private String lName;
-    private int age;
-    private String image;
+
 
     public User(int id) {
         this.id = id;
@@ -29,26 +28,7 @@ public abstract class User {
         return this.lName;
     }
 
-
-
     public void setfName(String fName) {this.fName = fName;}
 
     public void setLName(String lName) { this.lName = lName; }
 
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-}
